requirements.txt
streamlit
youtube-transcript-api
langchain
langchain-community
langchain-google-genai
faiss-cpu
sentence-transformers
python-dotenv
